pyproject.toml
[project]
name = "oc-projet7"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-api-python-client>=2.194.0",
    "google-auth>=2.49.2",
    "google-cloud-storage>=3.9.0",
    "pandas>=2.3.3",
    "polars>=1.36.1",
    "pymongo>=4.16.0",
]
